refactor(dependency_injection): Reorganize dependency injection files
